replace old training code
